spiffe-tls/pyproject.toml
[tool.poetry]
name = "spiffe-tls"
version = "0.1.0"
description = "TLS Support using SPIFFE"
authors = ["Max Lambrecht <maxlambrecht@gmail.com>"]
readme = "README.md"
license = "Apache-2.0"

packages = [
    { include = "spiffetls", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"  # >= 3.9, < 4.0
spiffe = { path = "../spiffe", develop = true }
pyOpenSSL = "^24.0"

[tool.poetry.dev-dependencies]
black = "^24.3"
mypy = "^1.9"
pytest = "^8.1"
pytest-mock = "^3.14"
pre-commit = "^3.7"
flake8 = "^7.0"
testutils = { path = "../testutils" }

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
files = ["src", "tests"]


[tool.black]
skip-string-normalization = true
target-version = ['py39']


[tool.pytest.ini_options]
addopts = [
    "--doctest-modules",
]
testpaths = [
    "tests",
]
